Extract structured preference category adapters
